+++ src/main/java/me/mgin/graves/inventory/Trinkets.java
@@ -1,13 +1,12 @@
 package me.mgin.graves.inventory;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import dev.emi.trinkets.TrinketSlot;
-import dev.emi.trinkets.api.SlotReference;
-import dev.emi.trinkets.api.TrinketComponent;
-import dev.emi.trinkets.api.TrinketInventory;
-import dev.emi.trinkets.api.TrinketsApi;
+import dev.emi.trinkets.api.*;
+import dev.emi.trinkets.api.TrinketEnums.DropRule;
 import me.mgin.graves.api.InventoriesApi;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.entity.player.PlayerEntity;
@@ -55,11 +54,13 @@ public class Trinkets implements InventoriesApi {
         Optional<TrinketComponent> component = TrinketsApi.getTrinketComponent(player);
         DefaultedList<ItemStack> itemStacks = DefaultedList.of();
 
-        component.ifPresent(trinketComponent -> trinketComponent.forEach((ref, itemStack) -> {
-            if (EnchantmentHelper.hasVanishingCurse(itemStack)) {
+        component.ifPresent(trinketComponent -> trinketComponent.forEach((ref, stack) -> {
+            DropRule rule = ref.inventory().getSlotType().getDropRule();
+
+            if (EnchantmentHelper.hasVanishingCurse(stack) || rule.equals(DropRule.KEEP)) {
                 itemStacks.add(ItemStack.EMPTY);
             } else {
-                itemStacks.add(itemStack);
+                itemStacks.add(stack);
             }
         }));
 
@@ -78,17 +79,19 @@ public class Trinkets implements InventoriesApi {
         DefaultedList<ItemStack> unequipped = DefaultedList.of();
 
         for (int i = 0; i < inventory.size(); i++) {
+            ItemStack stack = inventory.get(i);
+            if (stack.isEmpty()) continue;
+
+            // Do nothing & let the item be deleted
+            if (EnchantmentHelper.hasVanishingCurse(stack)) continue;
+
             // Add item to be returned as unequipped
-            if (EnchantmentHelper.hasBindingCurse(inventory.get(i))) {
-                unequipped.add(inventory.get(i));
+            if (EnchantmentHelper.hasBindingCurse(stack)) {
+                unequipped.add(stack);
                 continue;
             }
 
-            // Do nothing & let the item be deleted
-            if (EnchantmentHelper.hasVanishingCurse(inventory.get(i)))
-                continue;
-
-            equipItem(inventory.get(i), player, i);
+            equipItem(stack, player, i);
         }
 
         return unequipped;
@@ -102,7 +105,10 @@ public class Trinkets implements InventoriesApi {
     public void clearInventory(PlayerEntity player) {
         Optional<TrinketComponent> component = TrinketsApi.getTrinketComponent(player);
 
-        component.ifPresent(trinketComponent -> trinketComponent.forEach((ref, itemStack) -> {
+        component.ifPresent(trinketComponent -> trinketComponent.forEach((ref, stack) -> {
+            DropRule rule = ref.inventory().getSlotType().getDropRule();
+            if (rule.equals(DropRule.KEEP)) return;
+
             TrinketInventory inventory = ref.inventory();
             inventory.setStack(ref.index(), ItemStack.EMPTY);
         }));