+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-New: Updated to Minecraft v1.19.2.
+## 1.19.2-3.0.3
+**Fixed**
+- Trinkets `drop-rule` compatibility: Graves only handle slots with their `drop_rule` set to `default` now.
+- Items with Curse of Vanishing and Curse of Binding will prioritize vanishing now.